backend/requirements.txt
# requirements.txt

# --- Core Frameworks ---
fastapi
uvicorn
python-dotenv
python-multipart

# --- LangChain & AI Libraries ---
langchain
langchain-community
langchain-huggingface
langchain-google-genai
google-generativeai
sentence-transformers

# --- Data Handling ---
PyPDF2
cassio
cassandra-driver